+++ backend/auth/auth/urls.py
@@ -17,12 +17,7 @@ Including another URLconf
 from django.contrib import admin
 from django.urls import path, include
 
-from personal.views import (
-    home_screen_view
-)
-
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', home_screen_view, name='home_page'),
     path('api/', include('users.urls'))
 ]
