--- backend/personal/views.py
@@ -1,6 +1,0 @@
-from django.shortcuts import render
-
-# # Create your views here.
-def home_screen_view(request):
-    print(request.headers)
-    return render(request, "base.html", {})
