--- backend/templates/base.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>this is the title</title>
-</head>
-<body>
-    this is the body
-    {% block content %}
-
-    {% endblock content %}
-</body>
-<footer>this is the footer</footer>
-</html>
